feat: split dragging controls
